Move `thread_role` related code into a separate file
